+++ tools/gen/print_impl_template
@@ -1,4 +1,4 @@
-/* This file was generated by generate-bolts.py */
+/* This file was generated by generate-wire.py */
 /* Do not modify this file! Modify the _csv file it was generated from. */
 
 #include "${options.header_filename}"
